refactor: Update formkit.config.ts, package.json, tailwind.config.ts, nuxt.config.ts, es.json, en.json, and tailwind.css
This commit updates various configuration files and language files. It adds a new file, formkit.config.ts, which defines the formkit configuration with a theme of 'genesis'. It also adds a new dependency, "@formkit/nuxt", to the package.json file. Additionally, it updates the tailwind.config.ts file to include a new path for a configuration file, vueform.config.ts. The nuxt.config.ts file is updated to include the "@formkit/nuxt" module as a dependency. Finally, the es.json and en.json language files are updated with new translations for the "bring", "send", "emailSuccess", and "emailError" keys. The tailwind.css file is also updated to remove the "mt-6" class from the "p" element.